feat: implement distributed queue
A new method, IsHeld, has been added to the Mutex interface and its implementation to check if the mutex lock is held or not.
The main feature addition is a distributed queue implementation using etcd which is included in queue.go. The additional queue struct methods support functionality such as producing and consuming messages from the queue, as well as committing a message. The newQueue function was added to help with the initialization of a new distributed queue.
This will be beneficial in applications necessitating distributed systems with queue-like behavior.