improve: Update Docker and service configurations
This commit makes several updates to the Docker configurations and service settings. The changes include modifying the network settings for improved container communications and updating the logging levels for better system monitoring.

The network 'pudding-infra-net' has been renamed to 'pudding-net', making it attachable so that it's more flexible for communication between containers.

The Kafka control center has been commented out, which might be due to it being unnecessary for the current setup.

The configurations for the consul-server and etcd-host have also been updated. For the consul-server, its image was updated to a more recent version, and for the etcd-host, it was assigned a container name and network.

A new volume, `etcd_data`, has been added for data persistence of etcd-host.

The logging settings in broker's config.yaml are changed to provide a more detailed output.

A stop task has also been added to the Makefile to allow for easier teardown of the system.

Lastly, there is a change in 'broker/config.yaml' to enable TLS.