improve: Update Docker and service configurations (#379)
This commit makes several updates to the Docker configurations and service settings. The changes include modifying the network settings for improved container communications and updating the logging levels for better system monitoring.

The network 'pudding-infra-net' has been renamed to 'pudding-net', making it attachable so that it's more flexible for communication between containers.